move setup_network to connection_handler
