+++ main/connection_handler.h
@@ -9,6 +9,7 @@
 #include "lwip/netdb.h"
 
 // Function declarations for connection handling
+void setup_network(esp_netif_t** netif);
 int receive_data(struct netbuf** firstNetBuf, bool isMuted, esp_netif_t* netif);
 int fill_buffer(bool* first_netbuf_processed, int* rc1,
                 struct netbuf* firstNetBuf, char** start, uint16_t* len);
